requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
fire==0.6.0
Flask==3.0.2
Flask-Login==0.6.3
fonttools==4.50.0
gunicorn==22.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.1
pdf2docx==0.5.8
pillow==10.2.0
pymongo==4.8.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pypdf==4.1.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
termcolor==2.4.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
